# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-ready blog backend API for Alfafaa Community - a Muslim community blogging platform.",
        "title": "Alfafaa Blog API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Alfafaa Support",
            "url": "https://alfafaa.com/support",
            "email": "support@alfafaa.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/tags": {
            "get": {
                "description": "Get a paginated list of all tags",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "List tags",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by tag name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tags retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.ResponseWithMeta"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.TagResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new tag (requires editor role)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Create tag",
                "parameters": [
                    {
                        "description": "Tag data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.CreateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tag created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.TagResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - requires editor role",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "409": {
                        "description": "Tag already exists",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            }
        },
        "/tags/popular": {
            "get": {
                "description": "Get the most used tags sorted by usage count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Get popular tags",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of tags to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Popular tags retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.TagResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tags/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a tag (requires editor role)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Update tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tag update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.UpdateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.TagResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - requires editor role",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "404": {
                        "description": "Tag not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a tag (requires editor role)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Delete tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - requires editor role",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "404": {
                        "description": "Tag not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            }
        },
        "/tags/{slug}": {
            "get": {
                "description": "Get a single tag by its slug",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Get tag by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.TagResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid slug",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "404": {
                        "description": "Tag not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            }
        },
        "/tags/{slug}/articles": {
            "get": {
                "description": "Get a paginated list of articles with a specific tag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Get tag articles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.ResponseWithMeta"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.ArticleListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid slug",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "404": {
                        "description": "Tag not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_alfafaa_alfafaa-blog_internal_dto.ArticleListResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.PublicUserResponse"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.CategoryResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "featured_image_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "reading_time_minutes": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.TagResponse"
                    }
                },
                "title": {
                    "type": "string"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_alfafaa_alfafaa-blog_internal_dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "article_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_alfafaa_alfafaa-blog_internal_dto.CreateTagRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                }
            }
        },
...
        "github_com_alfafaa_alfafaa-blog_internal_dto.PublicUserResponse": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "profile_image_url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_alfafaa_alfafaa-blog_internal_dto.TagResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "usage_count": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_alfafaa_alfafaa-blog_internal_dto.UpdateTagRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2
                }
            }
        },
...
        "github_com_alfafaa_alfafaa-blog_internal_utils.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_alfafaa_alfafaa-blog_internal_utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_alfafaa_alfafaa-blog_internal_utils.ResponseWithMeta": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}